# src/main/resources/api.yaml
...
swagger: "2.0"
info:
  description: First API for the Tracey database models.
  version: 1.0.0
  title: Tracey_Database_API
  contact:
    email: fish_b@it.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /rikkygaray21/tracey-01/1.0.0
...
schemes:
- https
paths:
...
  /activities:
    get:
      summary: Get all activities.
      operationId: getAllActivityRecords
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Activities - Get collection - success
        "400":
          description: Activities - Get collection - failed
    post:
      summary: Add an activity
      operationId: postActivityToRecord
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Activities - Post collection - success
        "400":
          description: Activities - Post collection - failed
  /activities/{activityId}:
    get:
      summary: Get a specific activity by serching its Id
      operationId: getSingleActivityById
      produces:
      - application/json
      parameters:
      - name: activityId
        in: path
        description: String ID of the activity to get
        required: true
        type: string
      responses:
        "200":
          description: ActivitybyId - Get activity by its Id - success
        "400":
          description: ActivitybyId - Get activity by its Id - failed
    put:
      summary: Modify a specific activity by searching its Id
      operationId: putSingleActivityById
      produces:
      - application/json
      parameters:
      - name: activityId
        in: path
        description: String ID of the activity to put
        required: true
        type: string
      responses:
        "200":
          description: ActivitybyId - Put activity by its Id - success
        "400":
          description: ActivitybyId - Put activity by its Id - failed
    delete:
      summary: Delete a specific activity by searching its Id
      operationId: deleteSingleActivityById
      produces:
      - application/json
      parameters:
      - name: activityId
        in: path
        description: String ID of the activity to delete
        required: true
        type: string
      responses:
        "200":
          description: ActivitybyId - Delete activity by its Id - success
        "400":
          description: ActivitybyId - Delete activity by its Id - failed
    patch:
      summary: Modify a specific activity by serching its Id
      operationId: patchSingleActivityById
      produces:
      - application/json
      parameters:
      - name: activityId
        in: path
        description: String ID of the activity to patch
        required: true
        type: string
      responses:
        "200":
          description: ActivitybyId - Patch activity by its Id - success
        "400":
          description: ActivitybyId - Patch activity by its Id - failed
...